feat: agregar documentos para convenio
fix: correcciones menores

+++ backend/prisma/sql/conveniosGetPromedio.sql
@@ -16,7 +16,7 @@
 select 
     (
         avg(cur."notaFinal")
-    ):: numeric
+    )
     as "promedioPracticas"
 from "Convenio" con
 join "PTConvenio" ptc using ("idConvenio")
